Typescript the tooltips
